backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.8.2
pydantic-settings==2.4.0
PyJWT==2.8.0
SQLAlchemy==2.0.31
psycopg2-binary==2.9.9
python-multipart==0.0.9
alembic==1.13.2

